fix test placement
